# [SPARK-6953] [PySpark] speed up python tests