@rxin 2015-04-21
Note that this simply brings https://github.com/apache/spark/pull/5427 up to date.