@mengxr 2015-04-21
Sent you a PR at https://github.com/rxin/spark/pull/14.